+++ blocklets/ai-kit/api/src/hooks/pre-start.ts
@@ -2,6 +2,7 @@ import '@blocklet/sdk/lib/error-handler';
 
 import dotenv from 'dotenv-flow';
 
+import { ensureSqliteBinaryFile } from '../libs/ensure-sqlite';
 import logger from '../libs/logger';
 
 dotenv.config();
@@ -10,6 +11,10 @@ const { name } = require('../../../package.json');
 
 (async () => {
   try {
+    await ensureSqliteBinaryFile();
+
+    await import('../store/migrate').then((m) => m.default());
+
     process.exit(0);
   } catch (err) {
     logger.error(`${name} pre-start error`, err.message);