+++ blocklets/ai-kit/api/src/libs/env.ts
@@ -4,6 +4,9 @@ import Joi from 'joi';
 export default {
   ...env,
   chainHost: process.env.CHAIN_HOST || '',
-  openaiApiKey: process.env.OPENAI_API_KEY || '',
+  openaiApiKey: (process.env.OPENAI_API_KEY || '')
+    .split(',')
+    .map((i) => i.trim())
+    .filter(Boolean),
   verbose: Joi.boolean().validate(process.env.VERBOSE).value ?? false,
 };