+++ blocklets/ai-kit/api/src/routes/v1.ts
@@ -4,27 +4,20 @@ import { component } from '@blocklet/sdk/lib/middlewares';
 import { AxiosResponse } from 'axios';
 import { ParsedEvent, ReconnectInterval, createParser } from 'eventsource-parser';
 import { Request, Response, Router } from 'express';
+import { GPTTokens } from 'gpt-tokens';
 import Joi from 'joi';
 import {
   ChatCompletionRequestMessage,
-  Configuration,
   CreateEmbeddingRequest,
   CreateImageRequestResponseFormatEnum,
   CreateImageRequestSizeEnum,
-  OpenAIApi,
 } from 'openai';
 
+import { getAIProvider } from '../libs/ai-provider';
 import env from '../libs/env';
 import logger from '../libs/logger';
 import { ensureAdmin } from '../libs/security';
-
-function getAIProvider() {
-  const { openaiApiKey } = env;
-  if (!openaiApiKey) {
-    throw new Error('Missing required openai apiKey');
-  }
-  return new OpenAIApi(new Configuration({ apiKey: openaiApiKey }));
-}
+import Usage from '../store/models/usage';
 
 const router = Router();
 
@@ -36,13 +29,31 @@ async function status(_: Request, res: Response) {
 router.get('/status', ensureAdmin, status);
 router.get('/sdk/status', component.verifySig, status);
 
+export const Models = [
+  'gpt-3.5-turbo',
+  'gpt-3.5-turbo-0301',
+  'gpt-3.5-turbo-0613',
+  'gpt-3.5-turbo-16k',
+  'gpt-3.5-turbo-16k-0613',
+  'gpt-4',
+  'gpt-4-0314',
+  'gpt-4-0613',
+  'gpt-4-32k',
+  'gpt-4-32k-0314',
+  'gpt-4-32k-0613',
+] as const;
+
+export type Model = (typeof Models)[number];
+
 const completionsRequestSchema = Joi.object<
-  { stream?: boolean; model: string; temperature?: number } & (
+  { stream?: boolean; model: Model; temperature?: number } & (
     | { prompt: string; messages: undefined }
     | { prompt: undefined; messages: ChatCompletionRequestMessage[] }
   )
 >({
-  model: Joi.string().default('gpt-3.5-turbo'),
+  model: Joi.string()
+    .valid(...Models)
+    .default('gpt-3.5-turbo'),
   prompt: Joi.string(),
   messages: Joi.array()
     .items(
@@ -58,15 +69,17 @@ const completionsRequestSchema = Joi.object<
 }).xor('prompt', 'messages');
 
 async function completions(req: Request, res: Response) {
-  const { model, prompt, messages, stream, temperature } = await completionsRequestSchema.validateAsync(req.body, {
+  const { model, stream, temperature, ...input } = await completionsRequestSchema.validateAsync(req.body, {
     stripUnknown: true,
   });
 
   const openai = getAIProvider();
 
+  const messages = input.messages ?? [{ role: 'user', content: input.prompt }];
+
   const request: Parameters<typeof openai.createChatCompletion>[0] = {
     model,
-    messages: messages ?? [{ role: 'user', content: prompt }],
+    messages,
     stream,
     temperature,
   };
@@ -108,10 +121,27 @@ async function completions(req: Request, res: Response) {
 
     res.end();
   } else {
-    text = (await openai.createChatCompletion(request)).data.choices[0]?.message?.content.trim() ?? '';
+    const result = await openai.createChatCompletion(request);
+    text = result.data.choices[0]?.message?.content?.trim() ?? '';
 
     res.json({ text });
   }
+
+  const tokens = new GPTTokens({
+    model,
+    messages: messages
+      .concat({ role: 'assistant', content: text })
+      .filter(
+        (i): i is { role: GPTTokens['messages'][number]['role']; content: string } =>
+          i.role !== 'function' && Boolean(i.content)
+      ),
+  });
+
+  await Usage.create({
+    promptTokens: tokens.promptUsedTokens,
+    completionTokens: tokens.completionUsedTokens,
+    apiKey: openai.apiKey,
+  });
 
   if (env.verbose) logger.log('AI Kit completions output:', { text });
 }