+++ blocklets/ai-kit/api/src/store/migrations/20230823103001-create-usage.ts
@@ -1,0 +1,33 @@
+import { DataTypes } from 'sequelize';
+
+import type { Migration } from '../migrate';
+
+export const up: Migration = async ({ context: queryInterface }) => {
+  await queryInterface.createTable('Usages', {
+    id: {
+      type: DataTypes.STRING,
+      primaryKey: true,
+      allowNull: false,
+    },
+    createdAt: {
+      type: DataTypes.DATE,
+      allowNull: false,
+    },
+    updatedAt: {
+      type: DataTypes.DATE,
+      allowNull: false,
+    },
+    promptTokens: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+    },
+    completionTokens: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+    },
+  });
+};
+
+export const down: Migration = async ({ context: queryInterface }) => {
+  await queryInterface.dropTable('Usages');
+};