+++ blocklets/ai-kit/api/src/store/models/usage.ts
@@ -1,0 +1,53 @@
+import { CreationOptional, DataTypes, InferAttributes, InferCreationAttributes, Model } from 'sequelize';
+import { Worker } from 'snowflake-uuid';
+
+import { sequelize } from '../sequelize';
+
+const idGenerator = new Worker();
+
+const nextId = () => idGenerator.nextId().toString();
+
+export default class Usage extends Model<InferAttributes<Usage>, InferCreationAttributes<Usage>> {
+  declare id: CreationOptional<string>;
+
+  declare createdAt: CreationOptional<Date>;
+
+  declare updatedAt: CreationOptional<Date>;
+
+  declare promptTokens: number;
+
+  declare completionTokens: number;
+
+  declare apiKey?: string;
+}
+
+Usage.init(
+  {
+    id: {
+      type: DataTypes.STRING,
+      primaryKey: true,
+      allowNull: false,
+      defaultValue: nextId,
+    },
+    createdAt: {
+      type: DataTypes.DATE,
+      allowNull: false,
+    },
+    updatedAt: {
+      type: DataTypes.DATE,
+      allowNull: false,
+    },
+    promptTokens: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+    },
+    completionTokens: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+    },
+    apiKey: {
+      type: DataTypes.STRING,
+    },
+  },
+  { sequelize }
+);